backend/requirements.txt
# Core Django
Django>=4.2

# Gunicorn for production WSGI server
gunicorn>=21.0.0

# PostgreSQL database adapter
psycopg2-binary>=2.9.3

# Django REST Framework (API support)
djangorestframework>=3.14.0

# Django Simple JWT (for JWT Authentication)
djangorestframework-simplejwt>=5.3.0

# Optional: CORS headers (if frontend needs access)
django-cors-headers>=4.2.0

# Optional: Environment variables handling
python-dotenv>=1.0.0

# Celery (Task Queue)
celery>=5.3.4

# Redis (Broker for Celery)
redis>=5.0.1

# CORS Headers (if frontend needs access)
django-cors-headers>=4.2.0

# Environment Variables Handling
python-dotenv>=1.0.

pandas

numpy

Pillow

minio

django-storages

boto3

channels

daphne

channels-redis

django-cors-headers